requirements/requirements.txt
discord.py=2.1.1
